Add sh case in asm
Modify a halfword that doesn't contain the MSB or LSB. Perform a load immidiately after,
to check if the store has been through the store buffer, and reflects in the load operation.

t3 is being set to zero, initially. Hence, there is a problem if the load doesn't reflect.
I check this by branching to the end label if the register still contains the value zero.

Signed-off-by: B K Karthik <bkkarthik@pesu.pes.edu>