dcache_fill: change addi to add and load constants using li
addi takes a 12-bit immidiate value, but the value to be added here is a larger number. Use li to load the
constant into a register, as the compiler takes care of loading larger constants in li. Use register type
instruction to perform the addition.

Screw you Vishwa, I'm smart :p

Signed-off-by: B K Karthik <bkkarthik@pesu.pes.edu>