Set thrashing: Initialization of base address regs
Registers are initally initialzed using the logic that was used in
commit '09b34a0', However this logic has to be added with the address of
rvtest data. Hence once the addional registers that store the base
adresses are initialized they are added with the adress present in
register t2 (where the rvtest_data begins).

Signe-off-by: Vishweswaran K <vishwa.kans07@gmail.com>
gic: Initialization of base address regs

Registers are initally initialzed using the logic that was used in
commit '09b34a0', However this logic has to be added with the address of
rvtest data. Hence once the addional registers that store the base
adresses are initialized they are added with the adress present in
register t2 (where the rvtest_data begins).